requirements/requirements.txt
# alembic~=1.13.3
# asyncpg
# aiosqlite
fastapi~=0.115.0
pydantic-settings~=2.5.2
uvicorn~=0.32.0
SQLAlchemy~=2.0.36
SQLAlchemy-Utils~=0.41.2
# SQLAlchemy~=2.0.35
# fastapi-users[sqlalchemy]